configure: Add --without-libvirt option.
Normally we check if libvirt is installed, and the default is still to
do this.

However having this option makes it simpler to check if a change to
the code has broken the case where libvirt is not installed.

In fact, this change revealed one such place, which is also fixed in
this commit.