generator: Add 'blocking' flag to marking blocking (long-running) functions.
In languages like Python where we release a global lock around
long-running libguestfs functions, it is also useful to *not* release
this lock for small, non-blocking functions.

Therefore mark all functions with a 'blocking' boolean flag.  It
defaults to true, and is true by definition for all daemon functions.
For non-daemon functions, I have classified them manually.

Only when the blocking flag is set do we generate the code to release
and reacquire the lock around libguestfs calls.