launch: appliance: Really add dummy appliance slot (found by Coverity).
In the previous code, appliance was unconditionally set to NULL and so
we never added the dummy appliance slot to g->drives.  This probably
doesn't matter, since hotplugging doesn't work for the 'appliance'
attach-method.  Despite that, it's better to get this right.

This fixes commit ed7fda161e1f3d0beb02a368fcbcf5ed95dcdac1.