Refactor backends so they are modular.
Instead of hard-coding the list of backends into the core of
libguestfs in many places, make each backend into a true module which
registers itself with the core when the library is loaded.

This is a step towards making backends into loadable modules, but it
doesn't actually do that.

This is just code motion.