appliance: init: Don't write to /proc/sys/kernel/hotplug when starting udev (RHBZ#1020806).
/proc/sys/kernel/hotplug is an obsolete interface for processing
hotplug events.  It is not used by modern kernels, which should set
this path to the empty string (because they have
CONFIG_UEVENT_HELPER_PATH="").

Unfortunately our udev initialization code did:

  echo -e '\000\000\000\000' > /proc/sys/kernel/hotplug

which was intended to set this file to \0 bytes.  However on Debian,
where /bin/sh is "dash", not "bash", this writes "-e" to this file,
which breaks udev events.

Thanks: Peter Rajnoha for solving this.

https://bugzilla.redhat.com/show_bug.cgi?id=1020806#c10