v2v: Work around buggy `virsh dumpxml'.
The `virsh dumpxml' command doesn't work properly when the libvirt
source requires authentication.  This is because the authentication
prompts are sent to stdout, but stdout is also the place where we are
reading the output XML from.

Add a mini-binding to libvirt virDomainGetXMLDesc which avoids this,
getting the XML directly using the libvirt API.  Change existing
external calls to `virsh dumpxml' to use this API instead.