v2v: Remove lazy_refcounts option.
This seems* to cause disk corruption in the overlay files.  A
manifestation of this is that the qcow2 header is not written back
correctly to disk, resulting in the backing file of the overlay
"disappearing" and hence hardly any data being copied over to the
target.

* Note this could be for a variety of reasons and doesn't necessarily
indicate a fault in qemu or qcow2.