v2v: Fix code for editing grub2 console=... parameter (RHBZ#1143883).
When presented with a RHEL 7 guest, virt-v2v would print the following
warning message:

  virt-v2v: warning: could not update grub2 console: aug_get: no matching
  node (ignored)

This happened because on RHEL 7, /etc/sysconfig/grub is a symlink to
/etc/default/grub, and the Augeas grub2 lens returns entries under
/files/etc/default/grub/*, but the code was checking for
/files/etc/sysconfig/grub/*.

The fix is to check all the possible Augeas paths until one matches.
(Only the first match gets updated however).

Also, don't rebuild the grub2 config unless we've actually changed it.
This avoids doing a potentially failure-prone operation when it's not
necessary.

I suspect that the old virt-v2v code did not work correctly for RHEL 7
guests, but old virt-v2v didn't print a warning here, it just failed
to update silently.