v2v: Clean the RPM database early on (RHBZ#1143866).
If the RPM database in the source guest is corrupt, then subsequent
operations will fail randomly.

Old virt-v2v appears to have avoided this by cleaning the RPM database
very early on.  When translating the code I inadvertantly moved this
cleanup to a later stage -- in particular, after inspection of kernels
was done which requires running rpm commands.  Move this back to the
start.