resize: Rewrite the tests to use a stochastic testing method.
The previous tests were very limited, and ran the same two tests
on every run.

Randomly test:
 - MBR vs GPT
 - Primary, extended partitions
 - Expanding vs shrinking
 - Use of multiple --resize parameters
 - Source and target raw vs qcow2 formats
 - Resizing various different filesystem types
 - LV expand
 - extra partition or no extra partition

To run the test over and over again (useful for finding regressions in
virt-resize), do:

  while (cd resize; LIBGUESTFS_TRACE=1 ../run ./test-virt-resize.pl) >&/tmp/log ; do echo -n .; done

If it fails, you can see the failure in /tmp/log.

To rerun a specific test, search for the 'seed:' in the output of the
original test, and specify it on the command line:

  (cd resize; ../run ./test-virt-resize.pl --seed=<SEED>)