launch: libvirt: Use qemu-bridge-helper to implement a full network (RHBZ#1148012).
When using the libvirt backend, don't use the SLIRP.  Use
qemu-bridge-helper via libvirt to give us a full network connection.

One consequence of this is that 'ping' works in
'virt-builder --run-command'.

A less useful consequence is that the host firewall will prevent
connections on non-standard ports to the host.  So you can't (eg)
connect to a rsync daemon on the host listening on an arbitrary port,
which worked before.

The default bridge is 'virbr0', but you can override this by setting
LIBGUESTFS_BACKEND_SETTINGS=network_bridge=<some_bridge>

Note: this does not fix virt-rescue (since it overrides the default
backend and uses 'direct' for various reasons).