v2v: -i ova: Try hard to preserve the order of floppies and CDROMs from the original OVF.
Remove the common 'add_resource' function.

Duplicate that function into code for adding disks, and separate code
for adding removables (floppies & CDROMs).

The removables code should look for all removable devices
(ResourceType 14, 15, 16) in a single query, so that the order doesn't
depend on the order in which we do the queries.