p2v: gui: Get the correct button for cancel_button.
I was fetching cancel_button from the wrong dialog (conv_dlg, not
run_dlg), and therefore getting the wrong button.

This explains why the back button on the conversion dialog was always
insensitive.  It's because cancel_button was assigned to this button,
and then I was setting cancel_button to insensitive.