./run: Use 'prepend' function to build paths.
Add a bash function 'prepend' for intelligently prepending elements to
paths.  eg:

  prepend PYTHONPATH "/foo"

would set PYTHONPATH to "/foo" or "/foo:<previous-contents-of-PYTHONPATH>"

Tested by:

(1) Building and testing libguestfs twice: first without libguestfs
installed as a system library, and then with it installed.

(2) Examining the output of './run printenv' by hand and comparing
environment variables to the expected values.