v2v: test-harness: Match on subimages instead of screenshots.
Previously we fuzzy-matched on screenshots of the entire display.

There were several problems:

 - dates/times in images required fuzzy matching, but it's not clear
   how much fuzz to use (Windows 8.1/2012R2 timestamp is huge,
   requiring an enormous amount of fuzz)

- cannot cope with Windows 7 irregular placement of windows on
  the desktop

It's better to match on subimages.

This requires some changes to the test data, using gimp to clip the
images and remove any changing elements (dates/times/etc) but in the
long run it's more flexible than fuzzy matching on the whole display.

Note that because of peculiarities in 'compare' we still need to have
a little bit of fuzz in the matching.  Seemingly bit-for-bit identical
images still have a similarity > 0.