v2v: efi: Detect if the guest could boot with UEFI.
Use a heuristic to detect if the guest could boot with UEFI.

This is only used where we have missing metadata (in the
source.s_firmware == UnknownFirmware case).  Currently that only
applies for `-i disk' and `-i libvirt/libvirtxml'.

Eventually we'll be able to get this information from the libvirt
metadata (RHBZ#1217444), so it'll only be used for `-i disk'.