v2v: Support loading virtio-win drivers from virtio-win.iso (RHBZ#1234351).
This makes several changes to the handling of virtio-win drivers:

The VIRTIO_WIN_DIR environment variable has been renamed
VIRTIO_WIN (but you can still use the old name).

You can point the VIRTIO_WIN either at a RHEL virtio-win directory
(ie. /usr/share/virtio-win), OR at a loopback-mounted virtio-win ISO,
OR at the virtio-win.iso file itself.  In the latter case, libguestfs
is used to open the ISO file and read drivers from it.

The code is more flexible about the pathnames of drivers, because the
paths in the ISO are completely different from the paths in RHEL
/usr/share/virtio-win.