v2v: Introduce the concept of target buses.
The target VM will have several buses to which disks can be attached.
Commonly it will have an IDE bus, and possibly a virtio-blk "bus" (not
really a bus) and/or a SCSI bus.

Virt-v2v does not model this at the moment.  Disks are just added to
the output XML in the order that we write them, and so they can move
around with respect to the target VM.

This commit introduces the idea that we should model the target disk
buses, and we should try to assign fixed and removable disks to slots
on each of those buses.  In this commit, the modelling is not used by
any output mode, but that will be fixed in the next commit.