v2v: Pass target_buses to output object (RHBZ#1238053).
Pass the target_buses assignment to the output#create_metadata method.

Now output modes have a choice: they can either ignore the new
parameter and continue to use the flat list of targets.  This is
suitable for output modes that cannot model multiple buses (eg.
-o glance) or can model it but don't bother (currently -o rhev).

Or they can ignore the flat targets parameter and use the new
target_buses parameter, translating that into the appropriate list of
devices.  This is implemented in this commit for these modes:

 -o libvirt
 -o local
 -o qemu