p2v: tests: Don't fail when test machine has only a single hard disk.
virt-p2v looks in /sys/block to find disks, and ignores the hard disk
containing the root device.  This is correct when virt-p2v runs off
the ISO, but when running the tests on a machine that has a single
hard disk, all_disks would be NULL, resulting in a test failure.

Fix this by allowing all_disks to be NULL, but adding an extra check
later so that we don't perform the conversion with no hard disks (the
gui.c path already has this check, so it is only needed in the
kernel.c path).